[Net/Web] Fixed strict reduce types to 32 bit sizes, changed to memsize

+++ Common/Net/HTTPServer.cpp
@@ -353,7 +353,7 @@ void Server::HandleRequestDefault(const ServerRequest &request) {
 void Server::Handle404(const ServerRequest &request) {
 	INFO_LOG(IO, "No handler for '%s', falling back to 404.", request.resource());
 	const char *payload = "<html><body>404 not found</body></html>\r\n";
-	request.WriteHttpResponseHeader("1.0", 404, (int)strlen(payload));
+	request.WriteHttpResponseHeader("1.0", 404, strlen(payload));
 	request.Out()->Push(payload);
 }
