+++ Common/Net/Sinks.cpp
@@ -176,7 +176,7 @@ void InputSink::Fill() {
 		// Whatever isn't valid and follows write_ is what's available.
 		size_t avail = BUFFER_SIZE - std::max(write_, valid_);
 
-		int bytes = recv(fd_, buf_ + write_, (int)avail, MSG_NOSIGNAL);
+		int bytes = recv(fd_, buf_ + write_, avail, MSG_NOSIGNAL);
 		AccountFill(bytes);
 	}
 }
@@ -266,7 +266,7 @@ size_t OutputSink::PushAtMost(const char *buf, size_t bytes) {
 
 	if (valid_ == 0 && bytes > PRESSURE) {
 		// Special case for pushing larger buffers: let's try to send directly.
-		int sentBytes = send(fd_, buf, (int)bytes, MSG_NOSIGNAL);
+		int sentBytes = send(fd_, buf, bytes, MSG_NOSIGNAL);
 		// If it was 0 or EWOULDBLOCK, that's fine, we'll enqueue as we can.
 		if (sentBytes > 0) {
 			return sentBytes;
@@ -348,7 +348,7 @@ bool OutputSink::Flush(bool allowBlock) {
 	while (valid_ > 0) {
 		size_t avail = std::min(BUFFER_SIZE - read_, valid_);
 
-		int bytes = send(fd_, buf_ + read_, (int)avail, MSG_NOSIGNAL);
+		int bytes = send(fd_, buf_ + read_, avail, MSG_NOSIGNAL);
 #if !PPSSPP_PLATFORM(WINDOWS)
 		if (bytes == -1 && (errno == EAGAIN || errno == EWOULDBLOCK))
 			bytes = 0;
@@ -380,7 +380,7 @@ void OutputSink::Drain() {
 		// Let's just do contiguous valid.
 		size_t avail = std::min(BUFFER_SIZE - read_, valid_);
 
-		int bytes = send(fd_, buf_ + read_, (int)avail, MSG_NOSIGNAL);
+		int bytes = send(fd_, buf_ + read_, avail, MSG_NOSIGNAL);
 #if !PPSSPP_PLATFORM(WINDOWS)
 		if (bytes == -1 && (errno == EAGAIN || errno == EWOULDBLOCK))
 			bytes = 0;
